% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqdeg.R
\name{qqdeg}
\alias{qqdeg}
\title{qqdeg}
\usage{
qqdeg(file, object_type, group1, group2, fc_threshold = 1.5)
}
\arguments{
\item{file}{file in document}

\item{object_type}{gene data or te data}

\item{group1}{treat group}

\item{group2}{control group}

\item{fc_threshold}{foldchange value}
}
\value{
deg results and enrich results
}
\description{
qqdeg
}
\examples{
result <- qqdeg("rlim.xlsx","gene","male-ko","male-wt",fc_threshold = 1.5)
}
